data-service/requirements.txt
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1
requests==2.31.0
fredapi==0.6.1 